fix(build): resolve compilation errors and warnings in main loop and ledger
- Moved `PredictedStateDelta` and `ObservedStateDelta` structs to `include/types.h` to resolve visibility issues in `OutcomeGuard`.
- Removed redundant struct definitions from `include/shadow_executor.h` and `include/reality_sampler.h`.
- Fixed `C2374` redefinition error for `ctx` in `src/main.cpp`.
- Removed unused `ExecuteDecision` function in `src/main.cpp` to fix `C4505`.
- Updated `PredictionError` in `include/prediction_ledger.h` to use `double` instead of `int` to prevent precision loss warnings (`C4244`).
- Fixed unreferenced parameter warning in `src/shadow_executor.cpp`.

Files changed:
- include/types.h
- include/shadow_executor.h
- include/reality_sampler.h
- include/prediction_ledger.h
- src/main.cpp
- src/shadow_executor.cpp

+++ include/prediction_ledger.h
@@ -24,9 +24,9 @@
 #include "reality_sampler.h" // For ObservedStateDelta
 
 struct PredictionError {
-    int cpuError;
-    int thermalError;
-    int latencyError;
+    double cpuError;
+    double thermalError;
+    double latencyError;
 };
 
 class PredictionLedger {
